# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TASK MANAGER API",
        "version": "1.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "AppUser API"
                ],
                "summary": "Register",
                "operationId": "2ea304aede5bdc0b87244dd8b3d4b693",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "998943389449"
                                    },
                                    "fullname": {
                                        "type": "string",
                                        "example": "Sanjarbek"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Check user **token** and added new task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": null
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Qo'shildi"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "AppUser API"
                ],
                "summary": "Login",
                "operationId": "19a6e0a0a81a55a20deeb319128a167e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "phone": {
                                        "type": "string",
                                        "example": "998943389449"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Check user **token** and added new task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "id": 17,
                                                "phone": "998943389449",
                                                "fullname": "Sanjarbek",
                                                "token": "htg387htv73gw3yg8v9g8yrfg923f4",
                                                "password": "$2y$12$lkQctPgpsD2bYXkgHPJLt.DLpAJwjS.uXlCpc8Bt39pGv4G8tcNTK",
                                                "created_at": "2023-08-29T04:39:24.000000Z",
                                                "updated_at": null
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": "Xush Kelibsiz!"
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/get": {
            "get": {
                "tags": [
                    "AppUser API"
                ],
                "summary": "GetAppUser",
                "operationId": "250b7b81de871c693081714c961f73fa",
                "parameters": [
                    {
                        "name": "Token",
                        "in": "header",
                        "description": "User token",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check user **token** and added new task",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "success",
                                            "format": "boolean",
                                            "default": "true"
                                        },
                                        "data": {
                                            "description": "data",
                                            "format": "object",
                                            "example": {
                                                "id": 17,
                                                "phone": "998943389449",
                                                "fullname": "Sanjarbek",
                                                "token": "htg387htv73gw3yg8v9g8yrfg923f4",
                                                "password": "$2y$12$lkQctPgpsD2bYXkgHPJLt.DLpAJwjS.uXlCpc8Bt39pGv4G8tcNTK",
                                                "created_at": "2023-08-29T04:39:24.000000Z",
                                                "updated_at": null
                                            }
                                        },
                                        "message": {
                                            "description": "message",
                                            "format": "string",
                                            "default": ""
                                        },
                                        "error_code": {
                                            "description": "error_code",
                                            "format": "integer",
                                            "default": "0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}